Seperating files into folders and live updating lobby player list
